requirements.txt
unstructured[all-docs]
pillow
lxml
chromadb
tiktoken
langchain
langchain-community
langchain-openai
langchain-groq
langchain-chroma
python-dotenv
fastapi
uvicorn
python-multipart
